+++ mongodbreader/src/main/java/com/alibaba/datax/plugin/reader/mongodbreader/MongoDBReader.java
@@ -153,6 +153,12 @@ public class MongoDBReader extends Reader {
                         record.addColumn(new LongColumn((Integer) tempCol));
                     } else if (tempCol instanceof Long) {
                         record.addColumn(new LongColumn((Long) tempCol));
+                    } else  if (tempCol instanceof Document) {
+                        if (KeyConstant.isJsonType(column.getString(KeyConstant.COLUMN_TYPE))) {
+                            record.addColumn(new StringColumn(((Document) tempCol).toJson()));
+                        } else {
+                            record.addColumn(new StringColumn(tempCol.toString()));
+                        }
                     } else {
                         if (KeyConstant.isArrayType(column.getString(KeyConstant.COLUMN_TYPE))) {
                             String splitter = column.getString(KeyConstant.COLUMN_SPLITTER);